Implemented Systick
Systick didn't work because there was pc.printf in it. Removed timer because Derivative doesn't need it. The right and left designations were switched. Updated the encoder PID to reflect that so that they change the correct speedL or speedR.  Changed logic of turn functions. They needed && instead of ||. They will need work later on too though. Need to try to get Systick to be shorter. 10000 right now.